+++ src/bpp/models/cache.py
@@ -768,6 +768,7 @@ def rebuild_zwarte(offset=None, limit=None):
             "typ_kbn__skrot",
             "charakter_formalny__skrot",
             "charakter_formalny__charakter_sloty",
+            "oznaczenie_wydania",
         ],
     )